requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==2.3.7